World Reference Base for Soil Resources (4th Edition, 2022)

World Reference Base for Soil Resources (4th Edition, 2022)

Usage

data(WRB_4th_2022)

Format

An object of class list of length 3.

Details

A list containing three data.frame elements "rsg", "pq", and "sq" providing information on the 'Representative Soil Groups', 'Principal Qualifiers,' and 'Supplementary Qualifiers,' respectively.

Each element has the column "code" which is a number (1-32) referring to the position in the Reference Soil Groups, and the column "reference_soil_group" which is the corresponding group name. The "pq" and "sq" qualifier name columns (primary_qualifier and supplementary_qualifier) contain individual qualifier terms. Related qualifiers are identified using qualifier_group column derived from qualifier names separated with a forward slash " / "

The data.frame "rsg" has column "criteria", describing the logical criteria for each Reference Soil Group.

The data.frame "pq" has qualifier names in column "principal_qualifier"

The data.frame "sq" has column "supplementary_qualifier".

References

IUSS Working Group WRB. 2022. World Reference Base for Soil Resources. International soil classification system for naming soils and creating legends for soil maps. 4th edition. International Union of Soil Sciences (IUSS), Vienna, Austria.