Lookup Pre-calculated Soil Moisture Regimes by Taxon

Helper function for using latest version of the Keys to Soil Taxonomy standard lookup table for soil moisture information.

Usage

getTaxonSMR(
  taxon = NULL,
  code = NULL,
  as.is = FALSE,
  droplevels = FALSE,
  ordered = TRUE
)

Arguments

taxon

character. Vector of taxon names (order to subgroup level). These values are converted to taxon "codes"

code

character. Vector of taxon codes.

as.is

logical. Return character labels rather than an (ordered) factor? Default: FALSE

droplevels

logical. Drop unused levels? Default: FALSE

ordered

logical. Create an ordinal factor? Default: TRUE

Value

character or factor (when as.is=FALSE) containing soil moisture regime labels extracted from 13th edition Keys to Soil Taxonomy taxa using extractSMR()

See also

Examples

getTaxonSMR(c("aridisols", "haploxeralfs", NA, "abruptic durixeralfs", "ustic haplocryalfs"))
#>            aridisols         haploxeralfs                 <NA> 
#>      aridic (torric)                xeric                 <NA> 
#> abruptic durixeralfs   ustic haplocryalfs 
#>                xeric                ustic 
#> 7 Levels: aridic (torric) < ustic < xeric < udic < perudic < ... < peraquic