Get soil diagnostic horizons, characteristics and features

All parameters to this function are optional (default NULL). If specified, they are used as filters.

Usage

get_ST_features(
  group = NULL,
  chapter = NULL,
  name = NULL,
  page = NULL,
  multiline_sep = "\n",
  multiline_col = "criteria"
)

Arguments

group

optional filtering vector; one of: "Surface", "Subsurface", "Mineral", "Organic", "Mineral or Organic"

chapter

optional filtering vector for chapter number

name

optional filtering vector; these are the "names" of features used in headers

page

optional filtering vector; page number (12th Edition Keys to Soil Taxonomy)

multiline_sep

default "\n" returns multiline_col column as a character vector concatenated with "\n". Use NULL for list

multiline_col

character. vector of "multi-line" column names to concatenate. Default: "criteria"; use NULL for no concatenation.

Value

a subset of ST_features data.frame

Details

This is a wrapper method around the package data set ST_features.

See also

Examples

# get all features
str(get_ST_features())
#> 'data.frame':	84 obs. of  6 variables:
#>  $ group      : chr  "Soil Materials" "Soil Materials" "Soil Materials" "Soil Materials" ...
#>  $ name       : chr  "Mineral Soil Material" "Organic Soil Material" "Distinction Between Mineral Soils and Organic" "Soil Surface" ...
#>  $ chapter    : int  2 2 2 2 2 2 2 3 3 3 ...
#>  $ page       : int  3 3 3 4 4 4 4 7 8 8 ...
#>  $ description: chr  "Mineral soil material (less than 2.0 mm in diameter) either: 1. Is saturated with water for less than 30 days ("| __truncated__ "Soil material that contains more than the amounts of organic carbon described above for mineral soil material i"| __truncated__ "Soils Most soils are dominantly mineral material, but many mineral soils have horizons of organic material. For"| __truncated__ "The term \"soil surface\" is based on the upper limit of soil. The upper limit of soil is the boundary between "| __truncated__ ...
#>  $ criteria   : chr  "" "" "" "" ...

# get features in chapter 3
str(get_ST_features(chapter = 3))
#> 'data.frame':	77 obs. of  6 variables:
#>  $ group      : chr  "Surface" "Surface" "Surface" "Surface" ...
#>  $ name       : chr  "Anthropic Epipedon" "Folistic Epipedon" "Histic Epipedon" "Melanic Epipedon" ...
#>  $ chapter    : int  3 3 3 3 3 3 3 3 3 3 ...
#>  $ page       : int  7 8 8 8 9 9 10 10 11 11 ...
#>  $ description: chr  "The anthropic epipedon forms in human-altered or human- transported material (defined below). These epipedons f"| __truncated__ "" "" "" ...
#>  $ criteria   : chr  "Required Characteristics The anthropic epipedon consists of mineral soil material that shows evidence of the pu"| __truncated__ "Required Characteristics The folistic epipedon is a layer (one or more horizons) that is saturated for less tha"| __truncated__ "Required Characteristics The histic epipedon is a layer (one or more horizons) that is characterized by saturat"| __truncated__ "Required Characteristics The melanic epipedon has both of the following:\n1. An upper boundary at, or within 30"| __truncated__ ...

# get features on pages 18, 19, 20
get_ST_features(page = 18:20)
#>      group                                       name chapter page
#> 39 Mineral                       Anhydrous Conditions       3   18
#> 40 Mineral Coefficient of Linear Extensibility (COLE)       3   19
#> 41 Mineral                     Fragic Soil Properties       3   19
#> 42 Mineral                            Free Carbonates       3   19
#> 43 Mineral          Identifiable Secondary Carbonates       3   19
#> 44 Mineral          Interfingering of Albic Materials       3   19
#> 45 Mineral                                   Lamellae       3   20
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        description
#> 39                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Anhydrous (Gr. anydros, waterless) conditions refer to the moisture condition of soils in very cold deserts and other areas with permafrost (often dry permafrost). These soils typically have low precipitation (usually less than 50 mm water equivalent per year) and a moisture content of less than 3 percent by weight. Anhydrous soil conditions are similar to the aridic (torric) soil moisture regimes (defined below), except that the soil temperature at 50 cm is less than 5 o C throughout the year in the soil layers with these conditions.
#> 40                                                                                                                                                                                                                                                         The coefficient of linear extensibility (COLE) is the ratio of the difference between the moist length and dry length of a clod to its dry length. It is (Lm - Ld)/Ld, where Lm is the length at 33 kPa tension and Ld is the length when dry. COLE can be calculated from the differences in bulk density of the clod when moist and when dry. An estimate of COLE can be calculated in the field by measuring the distance between two pins in a clod of undisturbed soil at field capacity and again after the clod has dried. COLE does not apply if the shrinkage is irreversible. Durinodes Durinodes (L. durus, hard, and nodus, knot) are weakly cemented to indurated nodules or concretions with a diameter of 1 cm or more. The cement is SiO2 , presumably opal and microcrystalline forms of silica. Durinodes break down in hot concentrated KOH after treatment with HCl to remove carbonates but do not break down with concentrated HCl alone. Dry durinodes do not slake appreciably in water, but prolonged soaking can result in spalling of very thin platelets. Durinodes are firm or firmer and brittle when wet, both before and after treatment with acid. Some durinodes are roughly concentric when viewed in cross section, and concentric stringers of opal are visible under a hand lens.
#> 41 Fragic soil properties are the essential properties of a fragipan. They have neither the layer thickness nor volume requirements for the fragipan. Fragic soil properties are in subsurface horizons, although they can be at or near the surface in truncated soils. Aggregates with fragic soil properties have a firm or firmer rupture-resistance class and a brittle manner of failure when soil water is at or near field capacity. Air-dry fragments of the natural fabric, 5 to 10 cm in diameter, slake when they are submerged in water. Aggregates with fragic soil properties show evidence of pedogenesis, including one or more of the following: oriented clay within the matrix or on faces of peds, redoximorphic features within the matrix or on faces of peds, strong or moderate soil structure, and coatings of albic materials or uncoated silt and sand grains on faces of peds or in seams. Peds with these properties are considered to have fragic soil properties regardless of whether or not the density and brittleness are pedogenic. Soil aggregates with fragic soil properties must: 1. Show evidence of pedogenesis within the aggregates or, at a minimum, on the faces of the aggregates; and 2. Slake when air-dry fragments of the natural fabric, 5 to 10 cm in diameter, are submerged in water; and 3. Have a firm or firmer rupture-resistance class and a brittle manner of failure when soil water is at or near field capacity; and 4. Restrict the entry of roots into the matrix when soil water is at or near field capacity.
#> 42                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   The term "free carbonates" is used in the definitions of a number of taxa, is used as a criterion for the isotic mineralogy class, and is mentioned in the discussion of chemical analyses in the appendix. It refers to soil carbonates that are uncoated or unbound and that effervesce visibly or audibly when treated with cold, dilute HCl. The term "free carbonates" is nearly synonymous with the term "calcareous." Soils that have free carbonates generally have calcium carbonate as a common mineral, although sodium and magnesium carbonates are also included in this concept. Soils or horizons with free carbonates may have inherited the carbonate compounds from parent materials without any translocation or transformation processes acting on them. There is no implication of pedogenesis in the concept of free carbonates, as there is in identifiable secondary carbonates (defined below), although most forms of secondary carbonates are freely effervescent.
#> 43                                                                                                                                   The term "identifiable secondary carbonates" is used in the definitions of a number of taxa. It refers to translocated authigenic calcium carbonate that has been precipitated in place from the soil solution rather than inherited from a soil parent material, such as calcareous loess or limestone residuum. Identifiable secondary carbonates either may disrupt the soil structure or fabric, forming masses, nodules, concretions, or spheroidal aggregates (white eyes) that are soft and powdery when dry, or may be present as coatings in pores, on structural faces, or on the undersides of rock or pararock fragments. If present as coatings, the secondary carbonates cover a significant part of the surfaces. Commonly, they coat all of the surfaces to a thickness of 1 mm or more. If little calcium carbonate is present in the soil, however, the surfaces may be only partially coated. The coatings must be thick enough to be visible when moist. Some horizons are entirely engulfed by carbonates. The color of these horizons is largely determined by the carbonates. The carbonates in these horizons are within the concept of identifiable secondary carbonates. The filaments commonly seen in a dry calcareous horizon are within the meaning of identifiable secondary carbonates if the filaments are thick enough to be visible when the soil is moist. Filaments commonly branch on structural faces.
#> 44                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  The term "interfingering of albic materials" refers to albic materials that penetrate 5 cm or more into an underlying argillic, kandic, or natric horizon along vertical and, to a lesser degree, horizontal faces of peds. There need not be a continuous overlying albic horizon. The albic materials constitute less than 15 percent of the layer that they penetrate, but they form continuous skeletans (ped coatings of clean silt or sand defined by Brewer, 1976) 1 mm or more thick on the vertical faces of peds, which means a total width of 2 mm or more between abutting peds. Because quartz is such a common constituent of silt and sand, these skeletans are usually light gray when moist and nearly white when dry, but their color is determined in large part by the color of the sand or silt fraction.
#> 45                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Lamellae (lamella, if singular) are illuvial horizons less than 7.5 cm thick. Each lamella contains an accumulation of oriented silicate clay on or bridging sand and silt grains (and rock fragments if any are present). A lamella has more silicate clay than the overlying eluvial horizon.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           criteria
#> 39                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Required Characteristics Soils with anhydrous conditions have a mean annual soil temperature of 0 o C or colder. The layer from 10 to 70 cm below the soil surface has a soil temperature of less than 5o throughout the year and this layer:\n1. Includes no ice-impregnated permafrost; and\n2. Is dry (water held at 1500 kPa or more) in one-half or more of the soil for one-half or more of the time the layer has a soil temperature above 0 o C; or\nFigure 1. Soils that are plotted in the shaded area meet the andic soil properties criteria c, d, and e under item 3 of the required characteristics. To qualify as soils with andic properties, the soils must also meet the listed requirements for organic-carbon content, phosphate retention, and particle-size distribution.\n3. Has a rupture-resistance class of loose to slightly hard throughout when the soil temperature is 0 o C or colder, except where a cemented pedogenic horizon occurs.
#> 40                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 41                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 42                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 43                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> 44                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Required Characteristics Interfingering of albic materials is recognized if albic materials:\n1. Penetrate 5 cm or more into an underlying argillic, kandic, or natric horizon; and\n2. Are 2 mm or more thick between vertical faces of abutting peds; and\n3. Constitute less than 15 percent (by volume) of the layer that they penetrate.
#> 45 Required Characteristics A lamella is an illuvial horizon less than 7.5 cm thick formed in unconsolidated regolith more than 50 cm thick. Each lamella contains an accumulation of oriented silicate clay on or bridging the sand and silt grains (and rock fragments if any are present).\nEach lamella is required to have more silicate clay than the overlying eluvial horizon.\nLamellae occur in a vertical series of two or more, and each lamella must have an overlying eluvial horizon. (An eluvial horizon is not required above the uppermost lamella if the soil is truncated.)\nLamellae may meet the requirements for either a cambic or an argillic horizon. A combination of two or more lamellae 15 cm or more thick is a cambic horizon if the texture class is very fine sand, loamy very fine sand, or finer. A combination of two or more lamellae meets the requirements for an argillic horizon if there is 15 cm or more cumulative thickness of lamellae that are 0.5 cm or more thick and that have a clay content of either:\n1. 3 percent or more (absolute) higher than in the overlying eluvial horizon (e.g., 13 percent versus 10 percent) if any part of the eluvial horizon has less than 15 percent clay in the fine-earth fraction; or\n2. 20 percent or more (relative) higher than in the overlying eluvial horizon (e.g., 24 percent versus 20 percent) if all parts of the eluvial horizon have more than 15 percent clay in the fine- earth fraction.

# get the required characteristics for the mollic epipedon from list column
str(get_ST_features(name = "Mollic Epipedon")$criteria)
#>  chr "Required Characteristics The mollic epipedon consists of mineral soil material and, after mixing of the upper 1"| __truncated__